Do not log the missing bootstraper
This file can be missing during changes to the installed packages while it is being recreated.

The log entry here is pointless because if the following command succeeds, then this error is not actionable. If on the other hand the file could not be created, this will cause an error plus a log entry anyway.