Simplify the bit check and improve the JPG validation
The previous check was ported from the JS ExifUtil and is meant as a safety check.

An earlier version used a slightly different JPEG (that included a JFIF tag) but was replaced to some 30 bytes with this version.

This also means that any changes to the encoded image could break the following logic thus we convert it into `assert()` to fail hard if adjusted improperly.