Reset the peak memory usage for each job in BackgroundQueuePerformAction
This is specifically implemented in the controller instead of the
BackgroundQueueHandler, since the controller only has a single purpose:
Executing multiple different and a-priori unknown jobs. Measuring the peak
memory usage of the entire request is not useful there.

For other, user-facing, controllers the peak memory usage of an entire request
is more meaningful, since they also perform other tasks besides executing
background jobs. Thus we do not perform the reset in a generic location.