Disallow the use of `*` in `fromversion` without a preceding version number
This change bans the use of `fromversion="*"` which was meant for development purposes only but has slowly crept into production packages. It is an incredibly dangerous instructions that is not supported in update paths at all, for example, calculating the shortest update thread or the ability to skip in-between versions.

We’re officially supporting the use of wildcards with preceding version numbers like `fromversion="6.1.*"` for years. This is well supported, part of the actual spec and should be the only path forward with the exception of explicit version numbers, of course.