Move the notification out of the transaction
This change is motivated by https://www.woltlab.com/community/thread/311594-could-not-roll-back-transaction/ and merely aims to workaround whatever flaw is responsible for it.

This is a long-standing issues that only surfaces for very few self-hosted customers and has never been observed in our own environments.

The relevant code for the transaction handling had been throughly reviewed and tested multiple times without yielding any explanation for this issue.

At the end of the day this will most likely just make this error go away without changing the behavior of the code in the reactions.